@@ -15,12 +15,12 @@
     "hibp": "11.1.0",
     "jwt-decode": "^3.1.2",
     "query-string": "^8.1.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-day-picker": "^8.4.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-loader-spinner": "^5.3.4",
     "react-query": "^3.39.2",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.7.0",
     "react-toastify": "^9.1.1"
   },
   "scripts": {
@@ -57,13 +57,12 @@
   "devDependencies": {
     "@testing-library/cypress": "^9.0.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^28.1.6",
+    "@types/jest": "^29.2.6",
     "@types/node": "^16.18.11",
-    "@types/react": "^17.0.40",
-    "@types/react-dom": "^17.0.14",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router-dom": "^5.3.3",
     "@types/testing-library__cypress": "^5.0.9",
     "concurrently": "^7.6.0",
