@@ -1,19 +1,13 @@
+import { BrowserRouter } from "react-router-dom";
 import { useAuthContext } from "./hooks/context/auth";
 import { AuthenticationStatuses } from "./context/auth/types";
 import { Loading } from "./components/Loading";
 import { SignedOutHeader } from "./components/SignedOutHeader";
 import { SignedOutRouter } from "./components/SignedOutRouter";
-import { SignedInHeader } from "./components/SignedInHeader";
 import { SignedInRouter } from "./components/SignedInRouter";
-import { Footer } from "./components/Footer";
 
 export const App = () => {
-  const { authenticationStatus, getGroups, getFirstName, signOut } =
-    useAuthContext();
-
-  const handleSignout = async () => {
-    await signOut();
-  };
+  const { authenticationStatus } = useAuthContext();
 
   switch (authenticationStatus) {
     case AuthenticationStatuses.Initialising:
@@ -21,28 +15,12 @@ export const App = () => {
     case AuthenticationStatuses.NewPasswordRequired:
     case AuthenticationStatuses.NotSignedIn:
       return (
-        <>
+        <BrowserRouter>
           <SignedOutHeader />
           <SignedOutRouter authenticationStatus={authenticationStatus} />
-        </>
+        </BrowserRouter>
       );
     case AuthenticationStatuses.SignedIn:
-      const groups = getGroups();
-      const firstName = getFirstName();
-      return (
-        <>
-          <div className="app is-flex is-flex-direction-column">
-            <div className="is-flex-grow-1 is-flex-shrink-1">
-              <SignedInHeader
-                groups={groups}
-                firstName={firstName}
-                onSignout={handleSignout}
-              />
-              <SignedInRouter groups={groups} />
-            </div>
-            <Footer />
-          </div>
-        </>
-      );
+      return <SignedInRouter />;
   }
 };
