@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import { render, waitFor, screen } from "@testing-library/react";
-import { MemoryRouter } from "react-router-dom";
+import { createMemoryRouter, RouterProvider } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { AuthContextProvider } from "./context/auth";
 
@@ -8,11 +8,13 @@ export const renderInProvider = (component: ReactNode) => {
   const queryClient = new QueryClient();
 
   render(
-    <MemoryRouter>
-      <QueryClientProvider client={queryClient}>
-        <AuthContextProvider>{component}</AuthContextProvider>
-      </QueryClientProvider>
-    </MemoryRouter>
+    <QueryClientProvider client={queryClient}>
+      <AuthContextProvider>
+        <RouterProvider
+          router={createMemoryRouter([{ path: "*", element: component }])}
+        />
+      </AuthContextProvider>
+    </QueryClientProvider>
   );
 };
 
